rename the package prefix from com to org
